+++ src/client/store/data/slice.ts
@@ -4,9 +4,10 @@ import { Objects } from 'utils/objects'
 import { RecordAssessmentDatas } from 'meta/data'
 
 import { clearTableData } from './actions/clearTableData'
+import { createContact } from './actions/contacts/createContact'
+import { getContacts } from './actions/contacts/getContacts'
+import { updateContact } from './actions/contacts/updateContact'
 import { copyPreviousDatasources } from './actions/copyPreviousDatasources'
-import { createContact } from './actions/createContact'
-import { getContacts } from './actions/getContacts'
 import { getDescription } from './actions/getDescription'
 import { getLinkedDataSources } from './actions/getLinkedDataSources'
 import { getNodeValuesEstimations } from './actions/getNodeValuesEstimations'
@@ -15,7 +16,6 @@ import { getODPTableData } from './actions/getODPTableData'
 import { getTableData } from './actions/getTableData'
 import { postEstimate } from './actions/postEstimate'
 import { setNodeValues } from './actions/setNodeValues'
-import { updateContact } from './actions/updateContact'
 import { updateDescription } from './actions/updateDescription'
 import { updateNodeValues } from './actions/updateNodeValues'
 import { setNodeValuesReducer } from './extraReducers/setNodeValues'
